requirements.txt
beautifulsoup4==4.5.1
click==6.6
decorator==4.0.10
entrypoints==0.2.2
Flask==0.11.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.4
psycopg2==2.6.2
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.11.1
SQLAlchemy==1.0.15
Werkzeug==0.11.11
